requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
starlette==0.38.4


streamlit==1.38.0
requests==2.32.3
pandas==2.2.2
PyYAML==6.0.2